LayoutTests/http/wpt/webauthn/public-key-credential-create-success-hid.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.

PASS PublicKeyCredential's [[create]] with minimum options in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with authenticatorSelection { 'cross-platform' } in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with requireResidentKey { false } in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with userVerification { 'preferred' } in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with userVerification { 'discouraged' } in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with mixed options in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with two consecutive requests. 
PASS PublicKeyCredential's [[create]] with none attestation in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with direct attestation in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with indirect attestation in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with googleLegacyAppidSupport extension in a mock hid authenticator. 
PASS PublicKeyCredential's [[create]] with googleLegacyAppidSupport and appid extensions in a mock hid authenticator. 
